services/spark/requirements.txt
boto3
fiona==1.9.6
geopandas==0.13.2
kafka-python==2.0.2
numpy
opencv-python-headless
pillow
psycopg2-binary
pyproj==3.5.0
rasterio==1.3.10
segmentation-models-pytorch
shapely==2.0.6
timm
torch
torchvision
